requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
bottleneck=1.3.5=py311h5bb9823_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
bzip2=1.0.8=he774522_0
ca-certificates=2023.01.10=haa95532_0
contourpy=1.0.5=py311h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
glib=2.69.1=h5dc1a3c_2
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
icu=58.2=ha925a31_3
intel-openmp=2021.4.0=haa95532_3556
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.4=py311hd77b12b_0
krb5=1.19.4=h5b6d351_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_0
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.0=h6c2663c_2
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.10.3=h0ad7f3c_0
libxslt=1.1.37=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
matplotlib=3.7.1=py311haa95532_1
matplotlib-base=3.7.1=py311hf62ec03_1
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py311h2bbff1b_0
mkl_fft=1.3.1=py311h743a336_0
mkl_random=1.2.2=py311heda8569_0
munkres=1.1.4=py_0
numexpr=2.8.4=py311hffd1eac_0
numpy=1.24.3=py311ha6a8073_0
numpy-base=1.24.3=py311h877bb12_0
openssl=1.1.1t=h2bbff1b_0
packaging=23.0=py311haa95532_0
pandas=1.5.3=py311heda8569_0
pcre=8.45=hd77b12b_0
pillow=9.4.0=py311hd77b12b_0
pip=23.0.1=py311haa95532_0
ply=3.11=py311haa95532_0
pyparsing=3.0.9=py311haa95532_0
pyqt=5.15.7=py311hd77b12b_0
pyqt5-sip=12.11.0=py311hd77b12b_0
python=3.11.3=h966fe2a_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2022.7=py311haa95532_0
qt-main=5.15.2=he8e5bd7_8
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h2bbfb41_5
setuptools=66.0.0=py311haa95532_0
sip=6.6.2=py311hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h2bbff1b_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.2=py311h2bbff1b_0
tzdata=2023c=h04d1e81_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py311haa95532_0
xz=5.2.10=h8cc25b3_1
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
